Implémentation de la version d'essai du service

+++ JumboToRolls.API/Models/RollDefinition.cs
@@ -3,6 +3,6 @@ namespace JumboToRolls.API.Models;
 public class RollDefinition
 {
     public string SizeId { get; set; } = string.Empty;
-    public double MinDiameter { get; set; }
-    public double MaxDiameter { get; set; }
+    public int MinDiameter { get; set; }
+    public int MaxDiameter { get; set; }
 }
